Use io instead of codecs
This way we don't depend on a third party library, and also our code is more
compatible with Python 3.

Also removed uses of readlines(), it's no use since a file is already a line
iterator and readlines() loads everything in memory, which we definitely don't want
in this use case.